requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
numpy==1.20.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0
scikit-learn==0.24.1
scipy==1.6.1
sklearn==0.0
threadpoolctl==2.1.0
Werkzeug==1.0.1
